pyproject.toml
[tool.poetry]
name = "ai-api-factory"
version = "0.1.0"
description = "A project to connect multiple APIs for multi-agent applications"
authors = ["Anand Houston <anand@anandhouston.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
openai = "^1.33.0"
requests = "^2.32.3"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
pytest = "^8.2.2"
pydantic = "^2.7.3"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
